expense-tracker-backend/pyproject.toml
[tool.poetry]
name = "expense-tracker-backend"
version = "0.1.0"
description = ""
authors = ["Muhammad Abyaz <m.abyaz681@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.0"}
pydantic = "^2.8.2"
pymongo = "^4.8.0"
pymongoose = "^1.3.8"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
